analyser/requirements.txt
acpype
os_command_py
pdb_manip_py
pdb2pqr
msgpack
gromacs_py
urllib3
tqdm
threadpoolctl
suds
scipy
python-dateutil
plip
pillow
packaging
networkx
mrcfile
mmtf-python
kiwisolver
joblib
idna
gsd
fonttools
fasteners
cycler
contourpy
charset-normalizer
certifi
biopython
art
pdb-tools
requests
matplotlib
GridDataFormats
mdanalysis
pytz
tzdata
numpy==1.22.4
pandas==2.1.1
seaborn
scikit-learn
mysql-connector-python